+++ scriptsDB/table/3erNivelCreaciónTablas.sql
@@ -33,7 +33,7 @@ CREATE TABLE User ( -- Cuenta tipo usuario en Carlevix
     IdUser INT(10) NOT NULL, 
     IdMembership INT(10) NOT NULL,
     CVV INT(3) NOT NULL,
-    CardNumber INT(20) NOT NULL,
+    CardNumber BIGINT(20) NOT NULL,
     StartDateSus DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP, 
     EndDateSus DATETIME,
